airbnb-final/agent-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
alembic==1.13.3
python-dotenv==1.0.1
pymysql==1.1.1

# LangChain + tools
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5

httpx==0.27.2
python-dateutil==2.9.0.post0
